Move log_level_setter definition
