+++ cpp/include/raft/core/logger.hpp
@@ -22,17 +22,3 @@
 #if !defined(RAFT_COMPILED)
 #include "logger-inl.hpp"
 #endif
-
-namespace raft {
-struct log_level_setter {
-  explicit log_level_setter(int level)
-  {
-    prev_level_ = logger::get(RAFT_NAME).get_level();
-    logger::get(RAFT_NAME).set_level(level);
-  }
-  ~log_level_setter() { logger::get(RAFT_NAME).set_level(prev_level_); }
-
- private:
-  int prev_level_;
-};
-}  // namespace raft
